% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nodes_from_links.R
\name{get_nodes_from_links}
\alias{get_nodes_from_links}
\title{Get Nodes From Links}
\usage{
get_nodes_from_links(Links, source = "source", target = "target")
}
\arguments{
\item{Links}{A data frame containing links information.}

\item{source}{The name of the column in the Links data frame that contains the source nodes. Default is "source".}

\item{target}{The name of the column in the Links data frame that contains the target nodes. Default is "target".}
}
\value{
A data frame of unique nodes.
}
\description{
This function takes a data frame of links and creates the respective nodes dataframe.
}
\examples{
links <- data.frame(source = c("A", "B", "C"), target = c("B", "C", "A"))
get_nodes_from_links(links)
}
