% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankeyD3plus-package.R
\docType{package}
\name{sankeyD3plus-package}
\alias{sankeyD3plus}
\alias{sankeyD3plus-package}
\title{sankeyD3plus: D3 JavaScript Sankey diagrams in R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Creates 'D3' 'JavaScript' Sankey diagrams in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SchmidtPaul/sankeyD3plus}
  \item \url{https://schmidtpaul.github.io/sankeyD3plus/}
  \item Report bugs at \url{https://github.com/SchmidtPaul/sankeyD3plus/issues}
}

}
\author{
\strong{Maintainer}: Paul Schmidt \email{paul.schmidt@biomath.de} (https://schmidtpaul.github.io/)

Authors:
\itemize{
  \item Florian Breitwieser (github.com/fbreitwieser/sankeyD3)
  \item Christopher Gandrud
  \item J.J. Allaire
  \item Kenton Russell
  \item Mike Bostock (d3.js library, http://d3js.org) [copyright holder]
}

Other contributors:
\itemize{
  \item Alexander Tarashansky (github.com/atarashansky/sankeyD3/tree/master) [contributor]
  \item B.W. Lewis [contributor]
  \item Kevin Kuo [contributor]
  \item Charles Sese [contributor]
  \item Peter Ellis [contributor]
  \item Jonathan Owen [contributor]
  \item Jennifer Rogers [contributor]
}

}
\keyword{internal}
