% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveNetwork.R
\name{saveNetwork}
\alias{saveNetwork}
\title{Save a Sankey diagram to an HTML or PNG file}
\usage{
saveNetwork(
  network,
  file,
  selfcontained = TRUE,
  vwidth = network$width,
  vheight = network$height,
  zoom = 5,
  delay = 1
)
}
\arguments{
\item{network}{Network to save, i.e. result of \code{sankeyNetwork()}.}

\item{file}{file to be saved - either html or png.}

\item{selfcontained}{Whether to save the HTML as a single self-contained file
(with external resources base64 encoded) or a file with external resources
placed in an adjacent directory.}

\item{vwidth}{Viewport width in pixels. Defaults to network$width.}

\item{vheight}{Viewport height in pixels. Defaults to network$height.}

\item{zoom}{Zoom factor. 1 means actual size, 2 means 50\% zoom, etc.}

\item{delay}{Time to delay before taking webshot, in seconds.}
}
\description{
Save a Sankey diagram created via \code{sankeyNetwork()} to an HTML or PNG
file. The HTML can include its dependencies in an adjacent directory or can
bundle all dependencies into the HTML file (via base64 encoding). PNG files
are converted from HTML using the webshot package.
}
